+++ src/main/java/com/github/sweintritt/jmus/Application.java
@@ -8,50 +8,62 @@ import java.util.concurrent.CompletableFuture;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.Strings;
 import org.jline.terminal.Terminal;
 import org.jline.terminal.TerminalBuilder;
 import org.jline.utils.*;
-import uk.co.caprica.vlcj.factory.MediaPlayerFactory;
-import uk.co.caprica.vlcj.player.base.MediaPlayer;
-import uk.co.caprica.vlcj.player.base.MediaPlayerEventAdapter;
 
 @Slf4j
 @Getter
 @Setter
-public final class Application {
+final class Application {
 
-    private enum Mode {
-        SORTED,
-        RANDOM
+    public enum Mode {
+        HELP, ENTRIES
     }
 
-    private static final String STATUS = "[ jmus %s | %d files | vol:%d | %s ] (q)uit, (s)top, (p)lay, (b)ack, " +
-            "(n)ext, (+)vol, (-)vol, (r)and";
-    private static final MediaPlayerFactory MEDIA_FACTORY = new MediaPlayerFactory();
+    public enum Order {
+        SORTED, RANDOM
+    }
+
+    private static final String STATUS = "[ jmus %s | %d files | vol:%d | %s ] press h for help";
 
     private final List<Entry> entries = new LinkedList<>();
-    private final Queue<Entry> playStack = new LimitedLiFoQueue<>(100);
+    private final Queue<Entry> playStack = new LimitedStack<>(100);
+    private final Player player;
+    private final Terminal terminal;
+    private final Properties properties;
 
     private Entry entry;
-    private Mode mode = Mode.SORTED;
-    private String version;
-    private int volume = 50;
-    private MediaPlayer player;
-    private Terminal terminal;
+    private Mode mode = Mode.ENTRIES;
+    private Order order = Order.SORTED;
     private boolean running;
     private int index = -1;
 
-    public Application() throws IOException {
-       terminal = TerminalBuilder.builder().build();
+    Application() throws IOException {
+        this(new Player(), TerminalBuilder.builder().build());
+    }
+
+    Application(final Player player, final Terminal terminal) {
+        log.debug("init application");
+        this.terminal = terminal;
+        this.player = player;
+        this.properties = new Properties();
+
+        try {
+            properties.load(Application.class.getClassLoader().getResourceAsStream("application.properties"));
+        } catch (final IOException e) {
+            log.error("Unable to read version: {}", e.getMessage(), e);
+        }
+
+        log.debug("application ready");
     }
 
     /**
      * @param directory Root directory to scan for music
      */
-    public void run(final File directory) {
+    void run(final File directory) {
         try {
             log.info("scanning {}", directory.getAbsoluteFile());
             loadFiles(directory);
@@ -63,15 +75,14 @@ public final class Application {
             next();
             while (running) {
                 draw();
-                final int key = terminal.reader().read();
-                handleKey(key);
+                handleKey(terminal.reader().read());
             }
         } catch (final Exception e) {
             quit(e);
         }
     }
 
-    private void loadFiles(final File dir) {
+    void loadFiles(final File dir) {
         log.info("searching {}", dir.getName());
         final File[] files = dir.listFiles();
         if (files != null) {
@@ -86,39 +97,43 @@ public final class Application {
         }
     }
 
-    public void handleKey(final int key) {
+    void handleKey(final int key) {
         log.debug("key:{}", key);
         switch (key) {
             case 'n' -> next();
             case 'b' -> back();
-            case 'p' -> play();
-            case 's' -> stop();
+            case 'p' -> player.play();
+            case 's' -> player.pause();
             case 'q' -> quit();
-            case '+' -> setVolume(player.audio().volume() + 10);
-            case '-' -> setVolume(player.audio().volume() - 10);
+            case '+' -> player.setVolume(player.getVolume() + 10);
+            case '-' -> player.setVolume(player.getVolume() - 10);
             case 'r' -> toggleRandom();
+            case 'h' -> toggleHelp();
+            // TODO esc should also exit help view
             default -> log.trace("unknown key {}", key);
         }
     }
 
-    public void toggleRandom() {
-       if (Mode.RANDOM == mode) {
-           Collections.sort(entries);
-           mode = Mode.SORTED;
-       } else {
-           Collections.shuffle(entries);
-           mode = Mode.RANDOM;
-       }
-       index = entries.indexOf(entry);
+    void toggleHelp() {
+        if (Mode.HELP == mode) {
+            mode = Mode.ENTRIES;
+        } else {
+            mode = Mode.HELP;
+        }
     }
 
-    public void setVolume(final int volume) {
-        log.debug("set volume to {}", volume);
-        this.volume = Math.clamp(volume, 0, 100);
-        Optional.ofNullable(player).ifPresent(p -> p.audio().setVolume(this.volume));
+    void toggleRandom() {
+        if (Order.RANDOM == order) {
+            Collections.sort(entries);
+            order = Order.SORTED;
+        } else {
+            Collections.shuffle(entries);
+            order = Order.RANDOM;
+        }
+        index = entries.indexOf(entry);
     }
 
-    public void next() {
+    void next() {
         log.debug("playing next song");
         if (entry != null) {
             playStack.add(entry);
@@ -131,7 +146,7 @@ public final class Application {
     /**
      * Jump back one track in the list
      */
-    public void back() {
+    void back() {
         log.debug("playing previous song. index stack size: {}", playStack.size());
         if (!playStack.isEmpty()) {
             index = entries.indexOf(playStack.poll());
@@ -139,98 +154,57 @@ public final class Application {
         }
     }
 
-    public void play(final int index) {
-        Optional.ofNullable(player).ifPresent(p -> p.controls().stop());
-        Optional.ofNullable(player).ifPresent(MediaPlayer::release);
+    void play(final int index) {
+        player.stop();
         try {
             entry = entries.get(index);
             log.info("playing {}, index:{}", entry.getFile().getName(), index);
-            player = MEDIA_FACTORY.mediaPlayers().newMediaPlayer();
-            player.media().play(entry.getMedia().info().mrl());
-            player.events().addMediaPlayerEventListener(new MediaPlayerEventAdapter() {
-                @Override
-                public void finished(final MediaPlayer player) {
-                    next();
-                }
+            player.prepare(entry.getMedia().info().mrl());
+            player.onMediaEnd(() -> {
+                next();
+                draw();
             });
-            player.audio().setVolume(volume);
-            this.play();
+            player.play();
         } catch (final Exception e) {
             log.error("Error during playback: {} ", e.getMessage(), e);
         }
     }
 
-    public void stop() {
-        log.debug("stopping");
-        Optional.ofNullable(player).ifPresent(p -> p.controls().stop());
-    }
-
-    public void play() {
-        log.debug("playing");
-        Optional.ofNullable(player).ifPresent(p -> p.controls().start());
-    }
-
-    public void quit(final Exception e) {
+    void quit(final Exception e) {
         log.debug("quiting");
         setRunning(false);
-        stop();
+        player.stop();
+        player.release();
         clearScreen();
         if (e != null) {
             log.error(e.getMessage(), e);
             System.err.println("Error: " + e.getMessage());
+            System.exit(1);
         }
         System.exit(0);
     }
 
-    public void quit() {
+    void quit() {
         quit(null);
     }
 
-    public void clearScreen() {
+    void clearScreen() {
         log.debug("clear screen");
         terminal.puts(InfoCmp.Capability.clear_screen);
         terminal.flush();
     }
 
-    public void draw() {
+    void draw() {
         try {
             log.debug("draw");
             var rows = terminal.getHeight();
             var columns = terminal.getWidth();
             clearScreen();
 
-            if (entries.isEmpty()) {
-                for (int i = 0; i < rows; ++i) {
-                    terminal.writer().println(StringUtils.EMPTY);
-                }
-            } else {
-                final int index = entries.indexOf(entry);
-                // Try to position the current title in the middle of the screen
-                final int half = Math.floorDiv(rows, 2);
-                int startIndex = index - half;
-                if (index + half > entries.size()) {
-                    startIndex -= (index + half) - entries.size();
-                }
-
-                // Ensure that the start index is in bounds of the entry list
-                startIndex = Math.clamp(startIndex, 0, entries.size());
-                final int columnLength = Math.max(0, columns / 3);
-                log.debug("rows: {}, half: {}, index: {}, startIndex: {}, entries: {}", rows, half, index,
-                        startIndex, entries.size());
-                for (int i = startIndex; i < startIndex + rows - 1; ++i) {
-                    final Entry current = (i > entries.size() - 1) ? null : entries.get(i);
-
-                    if (current == null) {
-                        log.debug("no entry at {}", i);
-                        terminal.writer().println(StringUtils.EMPTY);
-                    } else if (i == index) {
-                        var styled = new AttributedString(getFullTitle(current, columnLength),
-                                AttributedStyle.DEFAULT.foreground(AttributedStyle.WHITE).background(AttributedStyle.BLUE));
-                        styled.println(terminal);
-                    } else {
-                        terminal.writer().println(getFullTitle(current, columnLength));
-                    }
-                }
+            switch (mode) {
+                case Mode.ENTRIES -> drawEntries(columns, rows);
+                case Mode.HELP -> drawHelp(rows);
+                default -> drawEmpty(rows);
             }
 
             // Print status line
@@ -243,36 +217,91 @@ public final class Application {
         }
     }
 
-    public String getStatusLine(final int length) {
-        var mode = (player.status().isPlaying()) ? "playing" : "stopped";
+    void drawHelp(final int rows) {
+        var help = List.of(
+                "jmus - v" + properties.get("version") + " - " + properties.get("license"),
+                StringUtils.EMPTY,
+                "--------------------------------------------------",
+                StringUtils.EMPTY,
+                "Simple audio player to play your local library.",
+                "jmus is designed to be very easy to use, with just",
+                "a few simple keys.",
+                StringUtils.EMPTY,
+                StringUtils.EMPTY,
+                "Mappings",
+                StringUtils.EMPTY,
+                " b: play previous song",
+                " h: show this help text",
+                " n: play next song",
+                " p: start playing",
+                " q: quit jmus",
+                " r: switch between random or sorted song order",
+                " s: stop playing",
+                " +: increase volume",
+                " -: decrease volume"
+        );
+
+        help.forEach(l -> terminal.writer().println(l));
+        for (var i = help.size(); i < rows - 1; ++i) {
+            terminal.writer().println(StringUtils.EMPTY);
+        }
+    }
+
+    void drawEntries(final int columns, final int rows) {
+        if (entries.isEmpty()) {
+            drawEmpty(rows);
+        } else {
+            // Try to position the current title in the middle of the screen
+            final int half = Math.floorDiv(rows, 2);
+            int startIndex = index - half;
+            if (index + half > entries.size()) {
+                startIndex -= (index + half) - entries.size();
+            }
+
+            // Ensure that the start index is in bounds of the entry list
+            startIndex = Math.clamp(startIndex, 0, entries.size());
+            final int columnLength = Math.max(0, columns / 3);
+            log.debug("rows: {}, half: {}, index: {}, startIndex: {}, entries: {}", rows, half, index,
+                    startIndex, entries.size());
+            for (int i = startIndex; i < startIndex + rows - 1; ++i) {
+                final Entry current = (i > entries.size() - 1) ? null : entries.get(i);
+
+                if (current == null) {
+                    log.debug("no entry at {}", i);
+                    terminal.writer().println(StringUtils.EMPTY);
+                } else if (i == index) {
+                    var styled = new AttributedString(getFullTitle(current, columnLength),
+                            AttributedStyle.DEFAULT.foreground(AttributedStyle.WHITE)
+                                    .background(AttributedStyle.BLUE));
+                    styled.println(terminal);
+                } else {
+                    terminal.writer().println(getFullTitle(current, columnLength));
+                }
+            }
+        }
+    }
+
+    void drawEmpty(final int rows) {
+        for (int i = 0; i < rows; ++i) {
+            terminal.writer().println(StringUtils.EMPTY);
+        }
+    }
+
+    String getStatusLine(final int length) {
         final String status = String.format(STATUS,
-                getVersion(),
+                properties.get("version"),
                 entries.size(),
-                Optional.ofNullable(player).map(p -> p.audio().volume()).orElse(0),
-                mode);
+                player.getVolume(),
+                player.isPlaying() ? "playing" : "stopped");
         return status + " ".repeat(Math.max(0, length - status.length()));
     }
 
-    public String getVersion() {
-        if (version == null) {
-            try {
-                version = "v"
-                        + new String(IOUtils.toByteArray(Objects.requireNonNull(this.getClass().getClassLoader()
-                                .getResourceAsStream("version.txt"))));
-            } catch (final IOException e) {
-                log.error("Unable to read version: {}", e.getMessage(), e);
-                version = StringUtils.EMPTY;
-            }
-        }
-        return version;
-    }
-
-    public String getFullTitle(final Entry entry, final int columnLength) {
+    String getFullTitle(final Entry entry, final int columnLength) {
         return fitToWidth(entry.getArtist(), columnLength) + fitToWidth(entry.getAlbum(), columnLength)
                 + fitToWidth(entry.getTitle(), columnLength);
     }
 
-    public String fitToWidth(final String message, final int width) {
+    String fitToWidth(final String message, final int width) {
         final String msg = StringUtils.trim(message);
         return StringUtils.abbreviate(StringUtils.trim(msg), "... ", width)
                 + StringUtils.SPACE.repeat(Math.max(0, width - StringUtils.length(msg)));