+++ Changelog.md
@@ -16,7 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added 'back' command.
-- Switch between random and sorted order of songs
+- Switch between random and sorted order of songs.
+- `h` no shows a help page.
 
 ### Fixed